requirements.txt
python-telegram-bot==21.3
yt-dlp
python-dotenv
urllib3
